+++ src/main/java/com/nl/rabobank/bankstatementvalidator/repository/BankStatementRepository.java
@@ -3,8 +3,8 @@ package com.nl.rabobank.bankstatementvalidator.repository;
 import com.nl.rabobank.bankstatementvalidator.domain.TransactionData;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BankStatementRepository extends CrudRepository<TransactionData, Integer> {
+public interface BankStatementRepository extends CrudRepository<TransactionData, Long> {
 
-    TransactionData findByReferenceNo(Integer referenceNo);
+    TransactionData findByReferenceNo(Long referenceNo);
 
 }
